Change Seq property in MetaData to ulong

+++ SparklerNet/Core/Model/MetaData.cs
@@ -9,7 +9,7 @@ namespace SparklerNet.Core.Model;
 public record MetaData
 {
     /// <summary>
-    ///     Whether the metric contains part of a multi-part message.
+    ///     Whether the metric contains part of a multipart message.
     /// </summary>
     public bool? IsMultiPart { get; init; }
 
@@ -24,9 +24,9 @@ public record MetaData
     public ulong? Size { get; init; }
 
     /// <summary>
-    ///     The sequence number of this part of a multipart metric. 0 ~ 255.
+    ///     The sequence number of this part of a multipart metric.
     /// </summary>
-    public int? Seq { get; init; }
+    public ulong? Seq { get; init; }
 
     /// <summary>
     ///     If this is a file metric, representing the file name.
